// src/types/protobuff/pb_lobby_mail.ts
export interface lobbyMailProp {
  propID: number;
  propCount: number;
}
// ...
export interface scMailGetprop {
  mailid: number;
  status: number;
  prop: lobbyMailProp[];
  curUserScore: number[];
}

export interface csMailSysmailid {
}

export interface scMailSysmailid {
  mailid: number[];
}

function createBaselobbyMailProp(): lobbyMailProp {
  return { propID: 0, propCount: 0 };
}

export const lobbyMailProp: MessageFns<lobbyMailProp> = {
  encode(message: lobbyMailProp, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.propID !== 0) {
      writer.uint32(8).int32(message.propID);
    }
    if (message.propCount !== 0) {
      writer.uint32(16).int64(message.propCount);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): lobbyMailProp {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaselobbyMailProp();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.propID = reader.int32();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.propCount = longToNumber(reader.int64());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): lobbyMailProp {
    return {
      propID: isSet(object.propID) ? globalThis.Number(object.propID) : 0,
      propCount: isSet(object.propCount) ? globalThis.Number(object.propCount) : 0,
    };
  },

  toJSON(message: lobbyMailProp): unknown {
    const obj: any = {};
    if (message.propID !== 0) {
      obj.propID = Math.round(message.propID);
    }
    if (message.propCount !== 0) {
      obj.propCount = Math.round(message.propCount);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<lobbyMailProp>, I>>(base?: I): lobbyMailProp {
    return lobbyMailProp.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<lobbyMailProp>, I>>(object: I): lobbyMailProp {
    const message = createBaselobbyMailProp();
    message.propID = object.propID ?? 0;
    message.propCount = object.propCount ?? 0;
    return message;
  },
};
// ...
export const scMailGetprop: MessageFns<scMailGetprop> = {
  encode(message: scMailGetprop, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.mailid !== 0) {
      writer.uint32(8).int64(message.mailid);
    }
    if (message.status !== 0) {
      writer.uint32(16).int32(message.status);
    }
    for (const v of message.prop) {
      lobbyMailProp.encode(v!, writer.uint32(26).fork()).join();
    }
    writer.uint32(34).fork();
    for (const v of message.curUserScore) {
      writer.int64(v);
    }
    writer.join();
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): scMailGetprop {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBasescMailGetprop();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.mailid = longToNumber(reader.int64());
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.status = reader.int32();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.prop.push(lobbyMailProp.decode(reader, reader.uint32()));
          continue;
        }
        case 4: {
          if (tag === 32) {
            message.curUserScore.push(longToNumber(reader.int64()));

            continue;
          }

          if (tag === 34) {
            const end2 = reader.uint32() + reader.pos;
            while (reader.pos < end2) {
              message.curUserScore.push(longToNumber(reader.int64()));
            }

            continue;
          }

          break;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): scMailGetprop {
    return {
      mailid: isSet(object.mailid) ? globalThis.Number(object.mailid) : 0,
      status: isSet(object.status) ? globalThis.Number(object.status) : 0,
      prop: globalThis.Array.isArray(object?.prop) ? object.prop.map((e: any) => lobbyMailProp.fromJSON(e)) : [],
      curUserScore: globalThis.Array.isArray(object?.curUserScore)
        ? object.curUserScore.map((e: any) => globalThis.Number(e))
        : [],
    };
  },

  toJSON(message: scMailGetprop): unknown {
    const obj: any = {};
    if (message.mailid !== 0) {
      obj.mailid = Math.round(message.mailid);
    }
    if (message.status !== 0) {
      obj.status = Math.round(message.status);
    }
    if (message.prop?.length) {
      obj.prop = message.prop.map((e) => lobbyMailProp.toJSON(e));
    }
    if (message.curUserScore?.length) {
      obj.curUserScore = message.curUserScore.map((e) => Math.round(e));
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<scMailGetprop>, I>>(base?: I): scMailGetprop {
    return scMailGetprop.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<scMailGetprop>, I>>(object: I): scMailGetprop {
    const message = createBasescMailGetprop();
    message.mailid = object.mailid ?? 0;
    message.status = object.status ?? 0;
    message.prop = object.prop?.map((e) => lobbyMailProp.fromPartial(e)) || [];
    message.curUserScore = object.curUserScore?.map((e) => e) || [];
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };
// ...
export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}